Make more robust

+++ scripts/setup.sh
@@ -14,7 +14,7 @@ cd $(dirname $0)/..
 [ $(basename $(pwd)) = 'svelte-starter' ] && exit 0
 
 # remove documentation
-rm -rf doc/
+[ -d doc ] && rm -rf doc/
 
 # substitute '{project-name}' with the current folder name
 sed -i '' 's/{project-name}/'"$(basename $(pwd))"'/g' package.json .env iframe-snippet.html
